+++ keyed-vals-hspec-tests/ChangeLog.md
@@ -2,8 +2,14 @@
 
 `keyed-vals-hspec-tests` uses [PVP Versioning][1].
 
+## 0.2.0.0 -- 2022-12-08
+
+* add tests that verify [KeyedVal.Handle.Typed][]
+* reduce the number of functions exposed for external use
+
 ## 0.1.0.0 -- 2022-11-28
 
 * Initial version.
 
 [1]: https://pvp.haskell.org
+[KeyedVals.Handle.Typed]: https://hackage.haskell.org/package/keyed-vals/docs/KeyedVals-Handle.Typed.html